Consistently use `alphaOrig` in TT cutoff logic. (#56)
Note that this does not change semantics, since the distinction only
matters in non-PV nodes, but the changed condition only matters in PV
nodes.

However, this way of writing it is more consistent and less confusing.

+++ src/chess-engine-lib/EvalT.h
@@ -56,5 +56,5 @@ inline constexpr EvalT kMateEval     = (EvalT)30'000;
 }
 
 [[nodiscard]] FORCE_INLINE constexpr EvalT clampNonMateEval(const int eval) {
-    return (EvalT)clamp((int)eval, -kMateEval + 1'000, kMateEval - 1'000);
+    return (EvalT)clamp(eval, -kMateEval + 1'000, kMateEval - 1'000);
 }
